More OO (see description for details)
Made the act of removing a tile (setting it to null in the board … NOT
a NullTile, literally using the java value null) a function on Boards.

Refactored the eliminatedTiles[][] attribute to only exist in an
EliminationBoard.

Calling removeTile in an EliminationBoard will call the Board version
of the function but will also properly updated the eliminatedTiles[][]
array.

Also, since it was easy to do at this point, implemented hasWon() on
EliminationBoards