game breaking bug fix
If you were selecting blocks, hitting one of the special buttons would
make it so that your selection would run infinitely.

This code makes it so that you can’t invoke any of the components in
our SpecialButtonsView when the game is not in a discrete state (i.e.
selection)